+++ openfisca_tools/microsimulation.py
@@ -395,3 +395,6 @@ class Microsimulation:
                 target, wrt=wrt, delta=delta, percent=percent
             )
         return df
+
+
+GeneralMicrosimulation = Microsimulation